refactor(onboarding): extract PasteDetectPanel as a shared, shell-agnostic core
Move the paste-textarea + Detect + candidate-list UI out of
onboarding/PasteAndDetect into a new reusable component at
components/paste-detect/PasteDetectPanel. It owns all internal state and
reuses the existing extractOnboardingCandidates service/types (no
duplication). Onboarding now imports the shared panel via a thin re-export,
keeping its import path and behavior identical.

No behavior change — this is the reuse foundation for surfacing Paste &
detect in the Connect AI picker.
